Added examples for linking Rust and Haskell
Linking is done through a C api (using the foreign function interface for both haskell and rust)

We compile the Haskell code as a shared object(.so) and expose the Haskell functions in the C api code. Then, we compile the C code as a shared library and use it in Rust. We use the build.rs script so all the files are compiled when we build and run the project with "cargo run". In said script, we also link the shared libraries and clean up unused files from the compilation.